WORKING FINAL VERSION 1
Fatto il refactoring dei folder components, css, img.
Varsione testata senza riscontrare bug

+++ php/socket/webSocketServer.php
@@ -7,11 +7,11 @@
  * Time: 19.58
  */
 
-require_once '../ajax/helper.php';
+require_once 'ajax/helper.php';
 
 use Ratchet\MessageComponentInterface;
 use Ratchet\ConnectionInterface;
-require_once '../database/connection.php';
+require_once 'database/connection.php';
 
 class webSocketServer implements MessageComponentInterface{
     protected $clients;
@@ -33,13 +33,16 @@ class webSocketServer implements MessageComponentInterface{
      * @throws \Exception
      */
     function onOpen(ConnectionInterface $conn){
-        error_log('UTENTE ' . $conn->resourceId . 'CONNESSO.');
+        error_log('UTENTE ' . $conn->resourceId . ' CONNESSO.');
+        error_log('SESSION STATUS: ' . session_status() );
         if (session_status() === PHP_SESSION_ACTIVE) {
             session_regenerate_id();
             error_log('SESSIONE RIGENERATA, ID SESSIONE: ' . session_id());
         }else if (session_status() !== PHP_SESSION_ACTIVE){
-            session_start();
-            error_log('SESSIONE CON ID ' . session_id() . ' INIZIATA');
+            if (session_status() !== PHP_SESSION_NONE) {
+                session_start();
+                error_log('SESSIONE CON ID ' . session_id() . ' INIZIATA');
+            }
         }else{
             error_log('IMPOSSIBILE APRIRE UNA SESSIONE');
         }
@@ -148,7 +151,157 @@ class webSocketServer implements MessageComponentInterface{
             case 'insert_location':{
                 $result['action'] = 'insert_location';
                 $query = $this->connection->insert_location($decoded_message['data']['user'], $decoded_message['data']['name'], $decoded_message['data']['description'],
-                    $decoded_message['data']['latitude'], $decoded_message['data']['longitude'], $decoded_message['data']['imageName']);
+                    $decoded_message['data']['latitude'], $decoded_message['data']['longitude'], $decoded_message['data']['imageName'], $decoded_message['data']['radius'],
+                    $decoded_message['data']['is_indoor']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'insert_floor':{
+                $result['action'] = 'insert_floor';
+                $query = $this->connection->insert_floor($decoded_message['data']['name'], $decoded_message['data']['map_image'], $decoded_message['data']['map_width'],
+                $decoded_message['data']['spacing'], $decoded_message['data']['location']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'save_drawing':{
+                $result['action'] = 'save_drawing';
+                $query = $this->connection->save_drawing($decoded_message['data']['lines'], $decoded_message['data']['floor']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'get_drawing':{
+                $result['action'] = 'get_drawing';
+                $query = $this->connection->get_drawing($decoded_message['data']['floor']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'get_all_types':{
+                $result['action'] = 'get_all_types';
+
+                $query = $this->connection->get_all_types();
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'get_tag_macs':{
+                $result['action'] = 'get_tag_macs';
+
+                $query = $this->connection->get_tag_macs($decoded_message['data']['tag']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'get_anchor_types':{
+                $result['action'] = 'get_anchor_types';
+
+                $query = $this->connection->get_anchor_types();
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'insert_mac':{
+                $result['action'] = 'insert_mac';
+
+                $query = $this->connection->insert_mac($decoded_message['data']['name'], $decoded_message['data']['type'], $decoded_message['data']['tag_id']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'insert_tag':{
+                $result['action'] = 'insert_tag';
+
+                $query = $this->connection->insert_tag($decoded_message['data']['name'], $decoded_message['data']['type'], $decoded_message['data']['macs']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'insert_anchor':{
+                $result['action'] = 'insert_anchor';
+
+                $query = $this->connection->insert_anchor($decoded_message['data']['name'], $decoded_message['data']['mac'], $decoded_message['data']['type'], $decoded_message['data']['ip'],
+                    $decoded_message['data']['rssi'], $decoded_message['data']['proximity'], $decoded_message['data']['permitteds'], $decoded_message['data']['neighbors'], $decoded_message['data']['floor']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'update_anchor_position':{
+                $result['action'] = 'update_anchor_position';
+
+                $query = $this->connection->update_anchor_position($decoded_message['data']['x'], $decoded_message['data']['y'], $decoded_message['data']['id']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'delete_anchor':{
+                $result['action'] = 'delete_anchor';
+
+                $query = $this->connection->delete_anchor($decoded_message['data']['anchor_id']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'delete_floor':{
+                $result['action'] = 'delete_floor';
+
+                $query = $this->connection->delete_floor($decoded_message['data']['floor_id']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'delete_tag':{
+                $result['action'] = 'delete_tag';
+
+                $query = $this->connection->delete_tag($decoded_message['data']['tag_id']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'delete_location':{
+                $result['action'] = 'delete_location';
+
+                $query = $this->connection->delete_location($decoded_message['data']['location_id']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'delete_mac':{
+                $result['action'] = 'delete_mac';
+
+                $query = $this->connection->delete_mac($decoded_message['data']['mac_id']);
 
                 ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
 
@@ -158,9 +311,23 @@ class webSocketServer implements MessageComponentInterface{
             case 'save_marker_image':{
                 $result['action'] = 'save_marker_image';
 
-                $decodedFile = explode('data:image/png;base64,', $decoded_message['data']['image']);
-                $decodedFile = base64_decode($decodedFile[1]);
-                $result['result'] = file_put_contents(MARKERS_IMAGES_PATH . $decoded_message['data']['imageName'], $decodedFile);
+                if (array_key_exists('image', $decoded_message['data'])) {
+                    $decodedFile = explode('data:image/png;base64,', $decoded_message['data']['image']);
+                    $decodedFile = base64_decode($decodedFile[1]);
+                    $result['result'] = file_put_contents(MARKERS_IMAGES_PATH . $decoded_message['data']['imageName'], $decodedFile);
+                }
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'save_floor_image':{
+                $result['action'] = 'save_floor_image';
+
+                if (array_key_exists('image', $decoded_message['data'])) {
+                    $decodedFile = explode('data:image/png;base64,', $decoded_message['data']['image']);
+                    $decodedFile = base64_decode($decodedFile[1]);
+                    $result['result'] = file_put_contents(FLOOR_IMAGES_PATH . $decoded_message['data']['imageName'], $decodedFile);
+                }
 
                 $this->clients[$from->resourceId]->send(json_encode($result));
                 break;
@@ -232,6 +399,26 @@ class webSocketServer implements MessageComponentInterface{
                 $this->clients[$from->resourceId]->send(json_encode($result));
                 break;
             }
+            case 'get_all_locations':{
+                $result['action'] = 'get_all_locations';
+
+                $query = $this->connection->get_all_locations();
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'get_locations_by_user':{
+                $result['action'] = 'get_location_by_user';
+
+                $query = $this->connection->get_locations_by_user($decoded_message['data']['user']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
             case 'get_floor_info':{
                 $result['action'] = 'get_floor_info';
                 $query = $this->connection->get_floor_info($decoded_message['data']['location'], $decoded_message['data']['floor']);
@@ -286,6 +473,15 @@ class webSocketServer implements MessageComponentInterface{
                 $this->clients[$from->resourceId]->send(json_encode($result));
                 break;
             }
+            case 'get_all_tags':{
+                $result['action'] = 'get_all_tags';
+                $query = $this->connection->get_all_tags();
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
             case 'get_tags_by_user':{
                 $result['action'] = 'get_tags_by_user';
                 $query = $this->connection->get_tags_by_user($decoded_message['data']['user']);
@@ -304,15 +500,15 @@ class webSocketServer implements MessageComponentInterface{
                 $this->clients[$from->resourceId]->send(json_encode($result));
                 break;
             }
-            case 'get_tags_by_location':{
-                $result['action'] = 'get_tags_by_location';
-                $query = $this->connection->get_tags_by_location($decoded_message['data']['location']);
-
-                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
-
-                $this->clients[$from->resourceId]->send(json_encode($result));
-                break;
-            }
+//            case 'get_tags_by_location':{
+//                $result['action'] = 'get_tags_by_location';
+//                $query = $this->connection->get_tags_by_location($decoded_message['data']['location']);
+//
+//                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+//
+//                $this->clients[$from->resourceId]->send(json_encode($result));
+//                break;
+//            }
             case 'get_events':{
                 $result['action'] = 'get_events';
                 $query = $this->connection->get_events();
@@ -340,9 +536,29 @@ class webSocketServer implements MessageComponentInterface{
                 $this->clients[$from->resourceId]->send(json_encode($result));
                 break;
             }
+            case 'change_mac_field':{
+                $result['action'] = 'change_mac_field';
+                $query = $this->connection->change_mac_field($decoded_message['data']['mac_id'], $decoded_message['data']['mac_field'],
+                                                            $decoded_message['data']['field_value']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
             case 'change_tag_field':{
                 $result['action'] = 'change_tag_field';
                 $query = $this->connection->change_tag_field($decoded_message['data']['tag_id'], $decoded_message['data']['tag_field'],
+                                                            $decoded_message['data']['field_value']);
+
+                ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
+
+                $this->clients[$from->resourceId]->send(json_encode($result));
+                break;
+            }
+            case 'change_location_field':{
+                $result['action'] = 'change_location_field';
+                $query = $this->connection->change_location_field($decoded_message['data']['location_id'], $decoded_message['data']['location_field'],
                                                             $decoded_message['data']['field_value']);
 
                 ($query instanceof db_errors) ? $result['result'] = $query->getErrorName() : $result['result'] = $query;
